+++ keras/src/ops/image.py
@@ -7,68 +7,6 @@ from keras.src.ops.operation import Operation
 from keras.src.ops.operation_utils import compute_conv_output_shape
 
 
-class Rot90(Operation):
-    def __init__(self, k=1, axes=(0, 1)):
-        super().__init__()
-        self.k = k
-        self.axes = axes
-
-    def call(self, array):
-        return backend.image.rot90(array, k=self.k, axes=self.axes)
-
-    def compute_output_spec(self, array):
-        array_shape = list(array.shape)
-        if len(array_shape) < 2:
-            raise ValueError(
-                "Input array must have at least 2 dimensions. "
-                f"Received: array.shape={array_shape}"
-            )
-        if len(self.axes) != 2 or self.axes[0] == self.axes[1]:
-            raise ValueError(
-                f"Invalid axes: {self.axes}. Axes must be a tuple of two different dimensions."
-            )
-        axis1, axis2 = self.axes
-        array_shape[axis1], array_shape[axis2] = array_shape[axis2], array_shape[axis1]
-        return KerasTensor(shape=array_shape, dtype=array.dtype)
-
-
-@keras_export("keras.ops.image.rot90")
-def rot90(array, k=1, axes=(0, 1)):
-    """Rotate an array by 90 degrees in the plane specified by axes.
-
-    This function rotates an array counterclockwise by 90 degrees `k` times
-    in the plane specified by `axes`. Supports arrays of two or more dimensions.
-
-    Args:
-        array: Input array to rotate.
-        k: Number of times the array is rotated by 90 degrees.
-        axes: A tuple of two integers specifying the plane for rotation.
-
-    Returns:
-        Rotated array.
-
-    Examples:
-
-    >>> import numpy as np
-    >>> from keras import ops
-    >>> m = np.array([[1, 2], [3, 4]])
-    >>> rotated = ops.image.rot90(m)
-    >>> rotated
-    array([[2, 4],
-           [1, 3]])
-
-    >>> m = np.arange(8).reshape((2, 2, 2))
-    >>> rotated = ops.image.rot90(m, k=1, axes=(1, 2))
-    >>> rotated
-    array([[[1, 3],
-            [0, 2]],
-           [[5, 7],
-            [4, 6]]])
-    """
-    if any_symbolic_tensors((array,)):
-        return Rot90(k=k, axes=axes).symbolic_call(array)
-    return backend.image.rot90(array, k=k, axes=axes)
-
 class RGBToGrayscale(Operation):
     def __init__(self, data_format=None):
         super().__init__()
