+++ keras/src/ops/numpy_test.py
@@ -19,6 +19,73 @@ from keras.src.ops import numpy as knp
 from keras.src.testing.test_utils import named_product
 
 
+class NumPyTestRot90(testing.TestCase):
+    def test_basic(self):
+        array = np.array([[1, 2], [3, 4]])
+        rotated = knp.rot90(array)
+        expected = np.array([[2, 4], [1, 3]])
+        assert np.array_equal(rotated, expected), f"Failed basic 2D test: {rotated}"
+
+    def test_multiple_k(self):
+        array = np.array([[1, 2], [3, 4]])
+
+        # k=2 (180 degrees rotation)
+        rotated = knp.rot90(array, k=2)
+        expected = np.array([[4, 3], [2, 1]])
+        assert np.array_equal(rotated, expected), f"Failed k=2 test: {rotated}"
+
+        # k=3 (270 degrees rotation)
+        rotated = knp.rot90(array, k=3)
+        expected = np.array([[3, 1], [4, 2]])
+        assert np.array_equal(rotated, expected), f"Failed k=3 test: {rotated}"
+
+        # k=4 (full rotation)
+        rotated = knp.rot90(array, k=4)
+        expected = array
+        assert np.array_equal(rotated, expected), f"Failed k=4 test: {rotated}"
+
+    def test_axes(self):
+        array = np.arange(8).reshape((2, 2, 2))
+        rotated = knp.rot90(array, k=1, axes=(1, 2))
+        expected = np.array([[[1, 3], [0, 2]], [[5, 7], [4, 6]]])
+        assert np.array_equal(rotated, expected), f"Failed custom axes test: {rotated}"
+
+    def test_single_image(self):
+        array = np.random.random((4, 4, 3))
+        rotated = knp.rot90(array, k=1, axes=(0, 1))
+        expected = np.rot90(array, k=1, axes=(0, 1))
+        assert np.allclose(rotated, expected), "Failed single image test"
+
+    def test_batch_images(self):
+        array = np.random.random((2, 4, 4, 3))
+        rotated = knp.rot90(array, k=1, axes=(1, 2))
+        expected = np.rot90(array, k=1, axes=(1, 2))
+        assert np.allclose(rotated, expected), "Failed batch images test"
+
+    def test_invalid_axes(self):
+        array = np.array([[1, 2], [3, 4]])
+        try:
+            knp.rot90(array, axes=(0, 0))
+        except ValueError as e:
+            assert (
+                "Invalid axes: (0, 0). Axes must be a tuple of two different dimensions."
+                in str(e)
+            ), f"Failed invalid axes test: {e}"
+        else:
+            raise AssertionError("Failed to raise error for invalid axes")
+
+    def test_invalid_rank(self):
+        array = np.array([1, 2, 3])  # 1D array
+        try:
+            knp.rot90(array)
+        except ValueError as e:
+            assert (
+                "Input array must have at least 2 dimensions." in str(e)
+            ), f"Failed invalid rank test: {e}"
+        else:
+            raise AssertionError("Failed to raise error for invalid input rank")
+
+
 class NumpyTwoInputOpsDynamicShapeTest(testing.TestCase):
     def test_add(self):
         x = KerasTensor((None, 3))
